Update some references to the forked repo
I found this fork today while looking to upgrade my eslint setup, and
finding the still not merged updates to support eslint and use the
latest major version of typescript-eslint on the
eslint-plugin-jsx-expressions repo.

Great that you've published an updated fork to npm! I just made some
updates here so that the readme in this repo refers to the correct
install instructions for the fork, and that the npm listing refers back
to this GitHub repo.